+++ lib/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 1
-/// Strings: 253
+/// Strings: 280
 ///
-/// Built on 2026-05-15 at 19:28 UTC
+/// Built on 2026-05-15 at 20:56 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import